Cleanup the code in ParkinbgHouse, changed properties in classes, created logic to find and park and also created function to write all values. See more.
Moved GetInput functions on the unique properties of the vehicles car, bus and motorcycle to the constructors.
Moved function GetRandomVehicle to VehicleBase class, made it a lot less code and easier to understand.
Added properties to Parkingspot.
Changed name from IVehiclaBase to IVehicle.
Changed properties in ParkingHouse and added a for-loop in constructor to create x amount of parkingspots.
Created logic for to find parkingspot for vehicle and parking it
Created a function to write out all values in ParkingHouse